[REFACT] API 응답 롤백
