+++ src/main/java/io/oeid/mogakgo/common/swagger/template/ChatSwagger.java
@@ -7,8 +7,8 @@ import io.oeid.mogakgo.common.base.CursorPaginationResult;
 import io.oeid.mogakgo.core.properties.swagger.error.SwaggerChatErrorExamples;
 import io.oeid.mogakgo.core.properties.swagger.error.SwaggerProjectErrorExamples;
 import io.oeid.mogakgo.core.properties.swagger.error.SwaggerUserErrorExamples;
+import io.oeid.mogakgo.domain.chat.application.dto.res.ChatRoomDataRes;
 import io.oeid.mogakgo.domain.chat.application.dto.res.ChatRoomPublicRes;
-import io.oeid.mogakgo.domain.chat.application.dto.res.ChatRoomRes;
 import io.oeid.mogakgo.domain.chat.presentation.dto.res.ChatDataApiRes;
 import io.oeid.mogakgo.domain.chat.presentation.dto.res.ChatRoomIdApiRes;
 import io.oeid.mogakgo.exception.dto.ErrorResponse;
@@ -43,7 +43,7 @@ public interface ChatSwagger {
         @Parameter(name = "pageSize", description = "요청할 데이터 크기", example = "5", required = true),
         @Parameter(name = "sortOrder", description = "정렬 방향", example = "ASC"),
     })
-    ResponseEntity<CursorPaginationResult<ChatRoomRes>> getChatRoomList(
+    ResponseEntity<CursorPaginationResult<ChatRoomPublicRes>> getChatRoomList(
         @Parameter(hidden = true) Long userId,
         @Parameter(hidden = true) CursorPaginationInfoReq pageable);
 
@@ -62,7 +62,7 @@ public interface ChatSwagger {
                 }
             ))
     })
-    ResponseEntity<ChatRoomPublicRes> getChatRoomDetailData(
+    ResponseEntity<ChatRoomDataRes> getChatRoomDetailData(
         @Parameter(hidden = true) Long userId,
         @Parameter(in = ParameterIn.PATH) String chatRoomId);
