+++ src/main/java/io/oeid/mogakgo/domain/chat/application/ChatService.java
@@ -1,7 +1,7 @@
 package io.oeid.mogakgo.domain.chat.application;
 
+import io.oeid.mogakgo.domain.chat.application.dto.res.ChatRoomDataRes;
 import io.oeid.mogakgo.domain.chat.application.dto.res.ChatRoomPublicRes;
-import io.oeid.mogakgo.domain.chat.application.dto.res.ChatRoomRes;
 import io.oeid.mogakgo.domain.chat.entity.document.ChatRoom;
 import io.oeid.mogakgo.domain.chat.entity.vo.ChatRoomDetail;
 import io.oeid.mogakgo.domain.chat.exception.ChatException;
@@ -44,18 +44,18 @@ public class ChatService {
         chatRoomRepository.save(chatRoom);
     }
 
-    public List<ChatRoomRes> findChatRoomsByUserId(Long userId, Long cursorId, int pageSize) {
+    public List<ChatRoomPublicRes> findChatRoomsByUserId(Long userId, Long cursorId, int pageSize) {
         return chatRoomRepository.findChatRoomsByUserId(userId, cursorId, pageSize).stream()
-            .map(ChatRoomRes::from).toList();
+            .map(chatRoom -> ChatRoomPublicRes.of(chatRoom, userId)).toList();
     }
 
-    public ChatRoomPublicRes findChatRoomDetailData(UUID roomId, Long userId) {
+    public ChatRoomDataRes findChatRoomDetailData(UUID roomId, Long userId) {
         var chatRoom = chatRoomRepository.findByRoomIdAndUserId(roomId, userId)
             .orElseThrow(() -> new ChatException(ErrorCode404.CHAT_ROOM_NOT_FOUND));
         var userInfo = chatRoom.getParticipants().values().stream()
             .filter(info -> !info.userId().equals(userId))
             .findFirst().orElseThrow(() -> new ChatException(ErrorCode404.CHAT_USER_NOT_FOUND));
-        return new ChatRoomPublicRes(chatRoom.getChatRoomDetail(), userInfo);
+        return new ChatRoomDataRes(chatRoom.getChatRoomDetail(), userInfo);
     }
 
     public UUID findChatRoomIdByProjectId(Long projectId) {
