+++ src/main/java/io/oeid/mogakgo/domain/chat/application/dto/res/ChatRoomPublicRes.java
@@ -1,38 +1,58 @@
 package io.oeid.mogakgo.domain.chat.application.dto.res;
 
 import io.oeid.mogakgo.domain.chat.application.vo.ChatUserInfoRes;
-import io.oeid.mogakgo.domain.chat.entity.vo.ChatRoomDetail;
-import io.oeid.mogakgo.domain.chat.entity.vo.ChatUserInfo;
+import io.oeid.mogakgo.domain.chat.entity.document.ChatRoom;
+import io.oeid.mogakgo.domain.chat.entity.enums.ChatStatus;
 import io.swagger.v3.oas.annotations.media.Schema;
 import java.time.LocalDateTime;
 import lombok.Getter;
-import lombok.NoArgsConstructor;
 
-@Schema(description = "채팅방 프로젝트 정보")
+@Schema(description = "채팅방 리스트 조회 응답")
 @Getter
-@NoArgsConstructor
 public class ChatRoomPublicRes {
 
-    @Schema(description = "프로젝트 설명")
-    private String meetDetail;
-    @Schema(description = "프로젝트 시작 시간")
-    private LocalDateTime meetStartTime;
-    @Schema(description = "프로젝트 위치 위도")
-    private Double meetLocationLatitude;
-    @Schema(description = "프로젝트 위치 경도")
-    private Double meetLocationLongitude;
-    @Schema(description = "프로젝트 종료 시간")
-    private LocalDateTime meetEndTime;
+    @Schema(description = "커서 ID")
+    private Long cursorId;
+    @Schema(description = "프로젝트 ID")
+    private Long projectId;
+    @Schema(description = "채팅방 ID")
+    private String chatRoomId;
+    @Schema(description = "마지막 메시지")
+    private String lastMessage;
+    @Schema(description = "마지막 메시지 생성 시간")
+    private LocalDateTime lastMessageCreatedAt;
+    @Schema(description = "채팅방 상태")
+    private ChatStatus status;
 
-    private ChatUserInfoRes chatUserInfoRes;
+    private final ChatUserInfoRes chatUserInfoRes;
 
-    public ChatRoomPublicRes(ChatRoomDetail chatRoomDetail, ChatUserInfo chatUserInfo) {
-        this.meetDetail = chatRoomDetail.getMeetDetail();
-        this.meetStartTime = chatRoomDetail.getMeetStartTime();
-        this.meetLocationLatitude = chatRoomDetail.getMeetLocationLatitude();
-        this.meetLocationLongitude = chatRoomDetail.getMeetLocationLongitude();
-        this.meetEndTime = chatRoomDetail.getMeetEndTime();
-        this.chatUserInfoRes = ChatUserInfoRes.from(chatUserInfo);
+    private ChatRoomPublicRes(Long cursorId, Long projectId, String chatRoomId, ChatStatus status,
+        ChatUserInfoRes chatUserInfoRes) {
+        this.status = status;
+        this.chatRoomId = chatRoomId;
+        this.cursorId = cursorId;
+        this.projectId = projectId;
+        this.chatUserInfoRes = chatUserInfoRes;
     }
 
+    private void addLastMessage(String lastMessage, LocalDateTime lastMessageCreatedAt) {
+        this.lastMessage = lastMessage;
+        this.lastMessageCreatedAt = lastMessageCreatedAt;
+    }
+
+    public static ChatRoomPublicRes of(ChatRoom chatRoom, Long userId) {
+        var chatUserInfo = chatRoom.getOpponentUserInfo(userId);
+        var chatRoomRes = new ChatRoomPublicRes(
+            chatRoom.getCursorId(),
+            chatRoom.getChatRoomDetail().getProjectId(),
+            chatRoom.getRoomId().toString(),
+            chatRoom.getChatStatus(),
+            ChatUserInfoRes.from(chatUserInfo)
+        );
+        var lastMessage = chatRoom.getLastMessage();
+        if (lastMessage != null) {
+            chatRoomRes.addLastMessage(lastMessage.getMessage(), lastMessage.getCreatedAt());
+        }
+        return chatRoomRes;
+    }
 }
